Moved PosixSubsystem, Error, Shell to own file
